refactor: split shell settings surface ownership
